refactor: Pagina [id] funciona botones + y -, qty de producto en carrito se mueve a navBar
